update std calc

+++ trainers/archer_trainer.py
@@ -596,16 +596,16 @@ class ArcherEval(EvalBatchRollout):
             'eval/avg_game_length': sum(game_length) / len(game_length) if game_length else 0,
             'eval/min_episode_reward': min(total_episode_scores) if total_episode_scores else 0,
             'eval/max_episode_reward': max(total_episode_scores) if total_episode_scores else 0,
-            'eval/std_episode_reward': torch.std(torch.tensor(total_episode_scores)).item() if total_episode_scores else 0,
+            'eval/std_episode_reward': torch.std(torch.tensor(total_episode_scores, dtype=torch.float32)).item() if total_episode_scores else 0,
             'eval/min_turn_reward': min(total_response_scores) if total_response_scores else 0,
             'eval/max_turn_reward': max(total_response_scores) if total_response_scores else 0,
-            'eval/std_turn_reward': torch.std(torch.tensor(total_response_scores)).item() if total_response_scores else 0,
+            'eval/std_turn_reward': torch.std(torch.tensor(total_response_scores, dtype=torch.float32)).item() if total_response_scores else 0,
             'eval/min_accumulated_reward': min(per_episode_response_sum) if per_episode_response_sum else 0,
             'eval/max_accumulated_reward': max(per_episode_response_sum) if per_episode_response_sum else 0,
-            'eval/std_accumulated_reward': torch.std(torch.tensor(per_episode_response_sum)).item() if per_episode_response_sum else 0,
+            'eval/std_accumulated_reward': torch.std(torch.tensor(per_episode_response_sum, dtype=torch.float32)).item() if per_episode_response_sum else 0,
             'eval/min_game_length': min(game_length) if game_length else 0,
             'eval/max_game_length': max(game_length) if game_length else 0,
-            'eval/std_game_length': torch.std(torch.tensor(game_length)).item() if game_length else 0,
+            'eval/std_game_length': torch.std(torch.tensor(game_length, dtype=torch.float32)).item() if game_length else 0,
         }
 
         return metrics
